vm: function call now follows the lua way
the `Call` instruction takes an address and a range of registers (2
registers). This range of registers is passed as arguments to the
function at `address`
the `Ret` instruction takes a range of registers. This range of
registers is written on the registers used as arguments. TODO: handle
more return arguments that input arguments